requirements.txt
streamlit
openai
python-docx
fpdf
python-pptx
